requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
geoip2==3.0.0
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.8
maxminddb==1.5.2
Pillow==7.0.0
pytube3==9.6.4
pytz==2019.3
redis==3.4.1
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
typing-extensions==3.7.4.1
urllib3==1.25.8
vine==1.3.0
zipp==3.1.0
